requirements.txt
discord.py
cffi
pynacl
